backend/requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.19.0
Flask==2.3.3
Werkzeug==2.3.7
flask-cors
